moved to user_login. should be successful tests
